source/requirements.txt
transformers
sentencepiece
protobuf
pillow